+++ packages/utils/src/genCopyable/index.tsx
@@ -34,9 +34,16 @@ export const genCopyable = (dom: React.ReactNode, item: any, text: string) => {
     const ellipsis =
       item.ellipsis && text
         ? {
-            tooltip: needTranText ? <div className="pro-table-tooltip-text">{dom}</div> : text,
+            tooltip:
+              // 支持一下 tooltip 的关闭
+              item?.tooltip !== false && needTranText ? (
+                <div className="pro-table-tooltip-text">{dom}</div>
+              ) : (
+                text
+              ),
           }
         : false;
+
     return (
       <Typography.Text
         style={{