+++ tests/form/drawerForm.test.tsx
@@ -6,13 +6,13 @@ import React from 'react';
 import { waitForComponentToPaint } from '../util';
 
 describe('DrawerForm', () => {
-  it('📦 trigger will simulate onVisibleChange', async () => {
+  it('📦 trigger will simulate onOpenChange', async () => {
     const fn = jest.fn();
     const wrapper = render(
       <DrawerForm
         width={600}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -32,7 +32,7 @@ describe('DrawerForm', () => {
       <DrawerForm
         width={600}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText
           name="name"
@@ -136,13 +136,13 @@ describe('DrawerForm', () => {
     expect(!!wrapper.baseElement.querySelector('input#test')).toBeFalsy();
   });
 
-  it('📦 drawer close button will simulate onVisibleChange', async () => {
+  it('📦 drawer close button will simulate onOpenChange', async () => {
     const fn = jest.fn();
     const wrapper = render(
       <DrawerForm
         visible
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -156,13 +156,13 @@ describe('DrawerForm', () => {
     expect(fn).toBeCalledWith(false);
   });
 
-  it('📦 drawer close button will simulate onVisibleChange', async () => {
+  it('📦 drawer close button will simulate onOpenChange', async () => {
     const fn = jest.fn();
     const wrapper = render(
       <DrawerForm
         visible
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -177,13 +177,13 @@ describe('DrawerForm', () => {
     expect(fn).toBeCalledWith(false);
   });
 
-  it('📦 reset button will simulate onVisibleChange', async () => {
+  it('📦 reset button will simulate onOpenChange', async () => {
     const fn = jest.fn();
     const wrapper = render(
       <DrawerForm
         visible
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -207,7 +207,7 @@ describe('DrawerForm', () => {
           onClose: () => fn(false),
         }}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -230,7 +230,7 @@ describe('DrawerForm', () => {
           onClose: () => fn(false),
         }}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -254,7 +254,7 @@ describe('DrawerForm', () => {
           onClose: () => onCloseFn(false),
         }}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -279,7 +279,7 @@ describe('DrawerForm', () => {
       <DrawerForm
         visible
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
         onFinish={async () => true}
       >
         <ProFormText name="name" />
@@ -302,7 +302,7 @@ describe('DrawerForm', () => {
       <DrawerForm
         visible
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,
@@ -335,7 +335,7 @@ describe('DrawerForm', () => {
           },
         }}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
+        onOpenChange={(open) => fn(open)}
       >
         <ProFormText name="name" />
       </DrawerForm>,