+++ packages/components/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.17](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.3.16...@ant-design/pro-components@2.3.17) (2022-10-13)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.3.16](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.3.15...@ant-design/pro-components@2.3.16) (2022-10-11)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.3.15](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.3.14...@ant-design/pro-components@2.3.15) (2022-10-11)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.3.14](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.3.13...@ant-design/pro-components@2.3.14) (2022-10-10)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.3.13](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.3.12...@ant-design/pro-components@2.3.13) (2022-09-28)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
 ## [2.3.12](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.3.11...@ant-design/pro-components@2.3.12) (2022-09-22)
 
 **Note:** Version bump only for package @ant-design/pro-components