+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.15](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.0.14...@ant-design/pro-descriptions@2.0.15) (2022-10-13)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [2.0.14](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.0.13...@ant-design/pro-descriptions@2.0.14) (2022-10-11)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [2.0.13](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.0.12...@ant-design/pro-descriptions@2.0.13) (2022-10-11)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [2.0.12](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.0.11...@ant-design/pro-descriptions@2.0.12) (2022-10-10)
+
+### Bug Fixes
+
+- **form:** 解决 moneySymbol 设置为 false 时金额格式化异常问题 ([#6012](https://github.com/ant-design/pro-components/issues/6012)) ([7f9001b](https://github.com/ant-design/pro-components/commit/7f9001b94be054a8766d0a954f64d84fd7e646a3))
+
+## [2.0.11](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.0.10...@ant-design/pro-descriptions@2.0.11) (2022-09-28)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [2.0.10](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.0.9...@ant-design/pro-descriptions@2.0.10) (2022-09-22)
 
 **Note:** Version bump only for package @ant-design/pro-descriptions