+++ packages/skeleton/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-skeleton@2.0.3...@ant-design/pro-skeleton@2.0.4) (2022-10-11)
+
+**Note:** Version bump only for package @ant-design/pro-skeleton
+
 ## [2.0.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-skeleton@2.0.2...@ant-design/pro-skeleton@2.0.3) (2022-09-09)
 
 ### Bug Fixes